3.4.1

+++ src/main/java/de/jeff_media/drop2inventory/handlers/PermissionChecker.java
@@ -52,42 +52,41 @@ public class PermissionChecker {
 
         // Permission
         if (!player.hasPermission(Permissions.ALLOW_USE)) {
-            main.debug("No permission for drop2inv");
+ if(main.isDebug()) main.debug("No permission for drop2inv");
             return false;
         }
 
         // Drop2Inventory enabled?
         if (!hasDrop2InvEnabled(player)) {
-            main.debug("Player has drop2inv disabled");
+ if(main.isDebug()) main.debug("Player has drop2inv disabled");
             return false;
         }
 
 
         // Disabled worlds
         if (main.isWorldDisabled(world.getName())) {
-            main.debug("World is disabled: " + world.getName());
+ if(main.isDebug()) main.debug("World is disabled: " + world.getName());
             return false;
         }
 
         // GameMode
         if (player.getGameMode() == GameMode.CREATIVE && !main.getConfig().getBoolean(Config.WORKS_IN_CREATIVE)) {
-            main.debug("Gamemode is creative");
+ if(main.isDebug()) main.debug("Gamemode is creative");
             return false;
         }
 
         if (dropReason == DropReason.BLOCK_BREAK) {
             if (!isAllowedForBlock(player, dropSubject.getBlock())) {
-                main.debug("block specific checks failed - aborting");
+ if(main.isDebug()) main.debug("block specific checks failed - aborting");
                 return false;
             }
         } else if (dropReason == DropReason.ENTITY_KILL) {
             if (!isAllowedForEntity(player, dropSubject.getEntity())) {
-                main.debug("entity specific checks failed - aborting");
+ if(main.isDebug()) main.debug("entity specific checks failed - aborting");
                 return false;
             }
         }
-
-        main.debug("drop collection is allowed");
+ if(main.isDebug()) main.debug("drop collection is allowed");
         return true;
     }
 
@@ -95,13 +94,13 @@ public class PermissionChecker {
 
         // Block drops enabled
         if (!main.getConfig().getBoolean(Config.COLLECT_BLOCK_DROPS)) {
-            main.debug("Block drops disabled");
+ if(main.isDebug()) main.debug("Block drops disabled");
             return false;
         }
 
         // Block Blacklist
         if (!main.getUtils().isBlockEnabled(block.getType())) {
-            main.debug("Block is not whitelisted / is blacklisted: " + block.getType());
+ if(main.isDebug()) main.debug("Block is not whitelisted / is blacklisted: " + block.getType());
             return false;
         }
 
@@ -110,7 +109,7 @@ public class PermissionChecker {
         Material itemInMainHandType = itemInMainHand == null ? null : itemInMainHand.getType();
         if (main.getConfig().getBoolean(Config.PERMISSIONS_PER_TOOL, false)
                 && !(Utils.hasPermissionForThisTool(itemInMainHandType, player))) {
-            main.debug("No permission for this tool: " + itemInMainHandType);
+ if(main.isDebug()) main.debug("No permission for this tool: " + itemInMainHandType);
             return false;
         }
 
@@ -125,12 +124,12 @@ public class PermissionChecker {
 
 
         if (entity.getLastDamageCause() == null || entity.getLastDamageCause().getCause() == null) {
-            main.debug("Could not get last Damage Cause");
+ if(main.isDebug()) main.debug("Could not get last Damage Cause");
         } else {
-            main.debug(entity.getLastDamageCause().getCause().name());
+ if(main.isDebug()) main.debug(entity.getLastDamageCause().getCause().name());
             if (main.getConfig().getBoolean(Config.IGNORE_DROPS_FROM_MOBS_KILLED_BY_LAVA)) {
                 if (entity.getLastDamageCause().getCause() == EntityDamageEvent.DamageCause.LAVA) {
-                    main.debug("ignore drops from mobs killed by lava: true");
+ if(main.isDebug()) main.debug("ignore drops from mobs killed by lava: true");
                     // TODO
                     //main.legacyDropDetectionManager.registerIgnoredLocation(event.getEntity().getLocation());
                     return false;
@@ -139,7 +138,7 @@ public class PermissionChecker {
 
             if (main.getConfig().getBoolean(Config.IGNORE_DROPS_FROM_MOBS_KILLED_BY_LAVA)) {
                 if (entity.getLastDamageCause().getCause() == EntityDamageEvent.DamageCause.HOT_FLOOR) {
-                    main.debug("ignore drops from mobs killed by magma: true");
+ if(main.isDebug()) main.debug("ignore drops from mobs killed by magma: true");
                     // TODO
                     //main.legacyDropDetectionManager.registerIgnoredLocation(event.getEntity().getLocation());
                     return false;
@@ -152,18 +151,18 @@ public class PermissionChecker {
             LivingEntity victim = (LivingEntity) entity;
 
             if (victim.getKiller() == null && victim.getType() != EntityType.ARMOR_STAND) {
-                main.debug("R: Killer is null");
+ if(main.isDebug()) main.debug("R: Killer is null");
                 return false;
             }
 
             if (!main.getUtils().isMobEnabled(victim)) {
-                main.debug("Mob is disabled: " + entity.getType().name());
+ if(main.isDebug()) main.debug("Mob is disabled: " + entity.getType().name());
                 return false;
             }
         }
 
         if (!main.getConfig().getBoolean(Config.COLLECT_MOB_DROPS)) {
-            main.debug("Collect mob drops is disabled");
+ if(main.isDebug()) main.debug("Collect mob drops is disabled");
             return false;
         }
